% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensimplex2-package.R
\docType{package}
\name{opensimplex2-package}
\alias{opensimplex2}
\alias{opensimplex2-package}
\title{opensimplex2: Generate Multi-Dimensional Open Simplex Noise}
\description{
Generate 2, 3 or 4-dimensional autocorrelated noise. The noise function is comparible to classic Perlin noise, but with less directional artifacts and lower computational overhead.
}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

Authors:
\itemize{
  \item Marco Ciaramella [copyright holder]
  \item KdotJPG [copyright holder]
}

}
\keyword{internal}
