% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{opensimplex_noise}
\alias{opensimplex_noise}
\title{TODO}
\usage{
opensimplex_noise(type = "S", width, height, depth, slice, frequency = 1)
}
\arguments{
\item{type}{TODO}

\item{width, height, depth, slice}{TODO}

\item{frequency}{TODO}
}
\value{
A \code{matrix} (in case of two dimensions) or \code{array} (in case of more
dimensions) of \code{numeric} values between -1 and +1. TODO
}
\description{
TODO
}
\examples{
mat <- opensimplex_noise("S", 100, 100)
image(mat)
}
