% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{simplex2d}
\alias{simplex2d}
\title{TODO}
\usage{
simplex2d(width, height, frequency = 1, seed = 0L)
}
\arguments{
\item{width, height}{TODO}

\item{frequency}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
mat <- simplex2d(100, 100)
image(mat)
}
