% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{simplex3d}
\alias{simplex3d}
\title{TODO}
\usage{
simplex3d(width, height, depth, frequency = 1, seed = 0L)
}
\arguments{
\item{width, height, depth}{TODO}

\item{frequency}{TODO}

\item{seed}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
mat <- simplex3d(100, 100, 100)
image(mat[,,1])
}
