% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{simplex4d}
\alias{simplex4d}
\title{TODO}
\usage{
simplex4d(width, height, depth, slice, frequency = 1)
}
\arguments{
\item{width, height, depth, slice}{TODO}

\item{frequency}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
mat <- simplex4d(25, 25, 25, 25)
image(mat[,,1,1])
}
