% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensimplex2-package.R
\docType{package}
\name{opensimplex2-package}
\alias{opensimplex2}
\alias{opensimplex2-package}
\title{opensimplex2: Generate Multi-Dimensional Open Simplex Noise}
\description{
Generate 2, 3 or 4-dimensional gradient noise. The noise function is comparable to classic Perlin noise, but with less directional artefacts and lower computational overhead. It can have applications in procedural generation or (flow fields) simulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pepijn-devries.github.io/opensimplex2/}
  \item \url{https://github.com/pepijn-devries/opensimplex2/}
  \item Report bugs at \url{https://github.com/pepijn-devries/opensimplex2/issues}
}

}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

Authors:
\itemize{
  \item Marco Ciaramella [copyright holder]
  \item KdotJPG [copyright holder]
}

}
\keyword{internal}
