% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{opensimplex_noise}
\alias{opensimplex_noise}
\title{TODO}
\usage{
opensimplex_noise(type = "F", width, height, depth, slice, frequency = 1)
}
\arguments{
\item{type}{TODO}

\item{width, height, depth, slice}{TODO}

\item{frequency}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
mat <- opensimplex_noise("F", 100, 100)
image(mat)
}
