% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{opensimplex_space}
\alias{opensimplex_space}
\title{TODO}
\usage{
opensimplex_space(type = "S", dimensions = 2L)
}
\arguments{
\item{type}{TODO}

\item{dimensions}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
